Move getParsedDateFromQueryString()
