+++ server/utils/utils.test.ts
@@ -1,3 +1,4 @@
+import { format } from 'date-fns'
 import {
   convertToTitleCase,
   getResultsPagingLinks,
@@ -7,6 +8,7 @@ import {
   properCase,
   sortByTimestamp,
   safeReturnUrl,
+  getParsedDateFromQueryString,
 } from './utils'
 import { getResultsPagingLinksTestData, sortByTimestampData } from './utils.testData'
 
@@ -123,6 +125,25 @@ describe('safeReturnUrl', () => {
   ].forEach(testData => {
     it(`should output ${testData.expected} when supplied with ${testData.input}`, () => {
       expect(safeReturnUrl(testData.input)).toBe(testData.expected)
+    })
+  })
+})
+
+describe('getParsedDateFromQueryString', () => {
+  const today = format(new Date(), 'yyyy-MM-dd')
+
+  ;[
+    {
+      input: '2022-05-22',
+      expected: '2022-05-22',
+    },
+    {
+      input: '2222-00-12',
+      expected: today,
+    },
+  ].forEach(testData => {
+    it(`should output ${testData.expected} when supplied with ${testData.input}`, () => {
+      expect(getParsedDateFromQueryString(testData.input)).toBe(testData.expected)
     })
   })
 })
